Fix mobile dashboard startup loading
